Add files via upload
Task description
The task for the stage zero involves storing and displaying the personal information of each team member using a specific data structure.

Task solution
For the task, I chose to use a tuple in Python using python fiddle as the code editor to store the information, which includes name, username, email, hobby, discipline, country, and preferred programming language. I created a tuple for each team member and stored them in a list, then used indexing to print the information in an organized format. I then saved the file code and then uploaded to GitHub. The code is stored in this repository.